+++ tests/testthat/test-utils.R
@@ -15,20 +15,33 @@ test_that("get_key_duplicates_util function", {
   expect_error(get_key_duplicates_util(df, c(1, 2, 3)))
   expect_error(get_key_duplicates_util(df, keys = c("a", "test")))
 
-  expect_silent(get_key_duplicates_util(df, keys))
-
+  # single duplicated value
   expect_true(dplyr::all_equal(
-    dplyr::tibble(a = factor("b", levels = c("a", "b", "c")), b = 3, rows = "3,4", n = 2L),
+    data.frame(a = factor("b", levels = c("a", "b", "c")), b = 3, rows = "3,4", n = 2L),
     get_key_duplicates_util(df, keys)
   ))
 
-  # Expect empty tibble if there are no duplicated key values
+
+  # multiple duplicates (and single key)
+  df <- as.data.frame(
+    list(a = c("a", "a", "b", "b", "c"), b = c(1, 2, 3, 3, 4), c = c(1, 2, 3, 4, 5)),
+    stringsAsFactors = TRUE
+  )
+  keys <- c("a")
+  expect_true(dplyr::all_equal(
+    data.frame(a = factor(c("a", "b"), levels = c("a", "b", "c")), rows = c("1,2", "3,4"), n = c(2L, 2L)),
+    get_key_duplicates_util(df, keys)
+  ))
+
+  # Expect empty data.frame if there are no duplicated key values
   df <- as.data.frame(
     list(a = c("a", "a", "b", "b", "c"), b = c(1, 2, 3, 4, 5), c = c(1, 2, 3, 4, 5)),
     stringsAsFactors = TRUE
   )
+  keys <- c("a", "b")
+
   expect_true(dplyr::all_equal(
-    dplyr::tibble(a = factor(x = NULL, levels = c("a", "b", "c")), b = double(0), rows = character(0), n = integer(0)),
+    data.frame(a = factor(x = NULL, levels = c("a", "b", "c")), b = double(0), rows = character(0), n = integer(0)),
     get_key_duplicates_util(df, keys)
   ))
 })