+++ NEWS.md
@@ -1,3 +1,8 @@
 # teal.data 0.0.0.9005
 
 * Initialize the package.
+
+Changes of functionality from when package was part of teal:
+
+* `get_key_duplicates` returns a `data.frame` instead of a `tibble`.
+* `dplyr` moved from imports to suggests.